+++ CHANGELOG.rst
@@ -5,9 +5,9 @@ Changelog
 ~~~~~~~~~~~~~~~~~~
 
 - Added conversion classmethods to Base:
-    - from_object
-    - from_mapping
-    - from_sequence
+    - from_object()
+    - from_mapping()
+    - from_sequence()
 
 0.0.4 (2016-07-11)
 ~~~~~~~~~~~~~~~~~~